+++ src/util.h
@@ -29,8 +29,8 @@ char *ag_strdup(const char *s);
 char *ag_strndup(const char *s, size_t size);
 
 typedef struct {
-    int start; /* Byte at which the match starts */
-    int end;   /* and where it ends */
+    size_t start; /* Byte at which the match starts */
+    size_t end;   /* and where it ends */
 } match;
 
 typedef struct {
@@ -59,7 +59,7 @@ const char *boyer_moore_strncasestr(const char *s, const char *find, const size_
 
 strncmp_fp get_strstr(enum case_behavior opts);
 
-int invert_matches(match matches[], int matches_len, const int buf_len);
+size_t invert_matches(match matches[], size_t matches_len, const size_t buf_len);
 void compile_study(pcre **re, pcre_extra **re_extra, char *q, const int pcre_opts, const int study_opts);
 
 void *decompress(const ag_compression_type zip_type, const void *buf, const int buf_len, const char *dir_full_path, int *new_buf_len);
